requirements.txt
fastapi
uvicorn
yfinance
requests
google-generativeai
python-dotenv